# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: longue-vue CMDB API
  version: 0.1.0
  summary: Kubernetes-centric CMDB aligned with the ANSSI SecNumCloud framework.
  description: |
    REST API for **longue-vue**, a CMDB for Kubernetes environments aligned with the
    ANSSI SecNumCloud (SNC) qualification framework.

    This version (0.1.0) is a walking skeleton covering:
      - Liveness and readiness probes
      - `Cluster` resource CRUD

    Subsequent versions will add Node, Namespace, Workload, Pod, and other
    Kubernetes entity kinds mapped onto the ANSSI cartography layers.

    Errors follow RFC 7807 (`application/problem+json`).

    Interactive API documentation is available at `/docs/` for authenticated
    operators; sign in via `/ui/login` first to use "Try it out".
servers:
  - url: /
    description: Current origin (recommended for browser-based "Try it out" via /docs/)
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /v1/workloads/{id}:
    parameters:
      - $ref: '#/components/parameters/WorkloadId'
    get:
      tags: [workloads]
      summary: Get a workload
      operationId: getWorkload
      security:
        - BearerAuth: [read]
      responses:
        '200':
          description: Workload details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workload'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [workloads]
      summary: Update mutable fields of a workload
      description: |
        Merge-patch semantics: fields omitted from the body are unchanged.
        `namespace_id`, `kind`, and `name` are immutable after creation.
      operationId: updateWorkload
      security:
        - BearerAuth: [write]
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/WorkloadUpdate'
      responses:
        '200':
          description: Updated workload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workload'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [workloads]
      summary: Delete a workload
      operationId: deleteWorkload
      security:
        - BearerAuth: [delete]
      responses:
        '204':
          description: Workload deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: |
        API token issued out-of-band. Send as `Authorization: Bearer <token>`.

        Scopes granted to a token:
          - `read`   — list and get operations
          - `write`  — create and update (implies `read` at the caller's discretion)
          - `delete` — remove resources
          - `admin`  — all of the above; reserved for future admin-only endpoints

        Tokens carrying the `admin` scope implicitly satisfy any scope required
        by an operation.

        Per ADR-0007, machine tokens are created by an admin in the UI
        (see `POST /v1/admin/tokens`) rather than injected through env vars.

    SessionCookie:
      type: apiKey
      in: cookie
      name: longue_vue_session
      description: |
        Human session cookie set by `POST /v1/auth/login`. HttpOnly,
        SameSite=Strict, 8-hour sliding expiry. Per ADR-0007.

        Every operation accepts either `BearerAuth` or `SessionCookie` —
        scope enforcement reads from the caller's role (for sessions) or
        token scopes (for bearer), so the two paths share the same
        downstream authorisation check.
  parameters:
...
    WorkloadId:
      name: id
      in: path
      required: true
      description: Server-assigned workload identifier.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Token is valid but lacks the scope required by the operation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    Layer:
      type: string
      description: |
        ANSSI cartography layer the asset belongs to. See ADR-0002 for the
        kind-to-layer mapping and the rationale behind the six-layer model.
      enum:
        - ecosystem
        - business
        - applicative
        - administration
        - infrastructure_logical
        - infrastructure_physical

    # ── Reconcile schemas (ADR-0009: push collector) ──────────────────
...
    ContainerList:
      type: array
      description: |
        Containers associated with the resource. For Pods this reflects the
        live runtime (name, image, image_id from containerStatuses, init
        flag); for Workloads it's the pod template's declared containers
        (name, image, init). Opaque in v1: each entry's shape is additive,
        so new collector-extracted fields don't require a spec bump.
      items:
        type: object
        additionalProperties: true
...
    WorkloadKind:
      type: string
      description: |
        Kubernetes workload controller kind. Casing matches the `kind` field
        Kubernetes itself emits. Additional kinds (Job, CronJob, ReplicaSet,
        ReplicationController) may be introduced as new enum values without a
        schema migration (see ADR-0003).
      enum:
        - Deployment
        - StatefulSet
        - DaemonSet
    WorkloadMutable:
      description: Fields on a Workload that clients may set and later update.
      type: object
      properties:
        replicas:
          type: integer
          minimum: 0
          description: Desired replica count. Nullable; DaemonSet has no scalar desired count.
        ready_replicas:
          type: integer
          minimum: 0
          description: Observed ready replica count. Nullable until first status update.
        containers:
          $ref: '#/components/schemas/ContainerList'
        labels:
          type: object
          description: Arbitrary user-supplied string key/value labels.
          additionalProperties:
            type: string
            maxLength: 256
          maxProperties: 64
        spec:
          type: object
          description: |
            Kind-specific fields the collector chooses to persist (e.g. strategy
            for Deployment, service_name for StatefulSet). Open-ended; stored
            as JSONB.
          additionalProperties: true
        containers_versions:
          type: object
          description: Map of container name to its latest-version info. Keys are absent when the image is not yet enriched (non-parseable tag, registry not in allowlist, or not yet processed).
          additionalProperties:
            $ref: '#/components/schemas/ContainerVersionInfo'
...
    WorkloadUpdate:
      description: |
        Merge-patch body. Only mutable fields; `namespace_id`, `kind`, and
        `name` are immutable. Omitted fields are left unchanged.
      allOf:
        - $ref: '#/components/schemas/WorkloadMutable'
        - type: object
          properties:
            application_id:
              type: string
              format: uuid
              nullable: true
              description: |
                Link this workload to an Application (ADR-0029). When both
                application_id and application_name are set, application_id wins.
                Setting to null unlinks (per merge-patch semantics, omit the field
                entirely to leave the link unchanged).
            application_name:
              type: string
              description: |
                Convenience: link by Application name; resolved server-side.
                See application_id.
    Workload:
      description: A Kubernetes workload controller registered in the CMDB.
      allOf:
        - $ref: '#/components/schemas/WorkloadMutable'
        - type: object
          required: [id, namespace_id, kind, name, layer, created_at, updated_at]
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            namespace_id:
              type: string
              format: uuid
            kind:
              $ref: '#/components/schemas/WorkloadKind'
            name:
              type: string
              pattern: '^[a-z0-9]([-a-z0-9.]*[a-z0-9])?$'
              minLength: 1
              maxLength: 253
            layer:
              allOf:
                - $ref: '#/components/schemas/Layer'
              readOnly: true
              description: Always `applicative` for Workload. Set by the server.
            namespace_name:
              type: string
              nullable: true
              readOnly: true
              description: Denormalized `namespaces.name`. Null on orphans. See ADR-0027.
            cluster_id:
              type: string
              format: uuid
              nullable: true
              readOnly: true
              description: Denormalized `namespaces.cluster_id`. Null on orphans. See ADR-0027.
            cluster_name:
              type: string
              nullable: true
              readOnly: true
              description: Denormalized `clusters.name`. Null on orphans. See ADR-0027.
            application_id:
              type: string
              format: uuid
              nullable: true
              description: "Link to an Application (ADR-0029); operator-curated. Null when unlinked."
            effective_dict:
              allOf:
                - $ref: '#/components/schemas/EffectiveDICT'
              readOnly: true
              description: |
                Read-only inherited DICT classification (ADR-0029 §6).
                Computed at read time with application-wins precedence; never
                accepted on PATCH/POST bodies.
            created_at:
              type: string
              format: date-time
              readOnly: true
            updated_at:
              type: string
              format: date-time
              readOnly: true
    EffectiveDICT:
      description: |
        Read-only, computed DICT classification for a Workload or VM
        (ADR-0029 §6). `source` records where the values were inherited from:
        `application` when the linked Application carries any DICT axis,
        `workload`/`namespace` when the entity's own classification is used,
        or `none` when neither is set.
      type: object
      required: [source]
      properties:
        disponibilite:
          type: integer
          minimum: 0
          maximum: 4
          nullable: true
          readOnly: true
        integrite:
          type: integer
          minimum: 0
          maximum: 4
          nullable: true
          readOnly: true
        confidentialite:
          type: integer
          minimum: 0
          maximum: 4
          nullable: true
          readOnly: true
        tracabilite:
          type: integer
          minimum: 0
          maximum: 4
          nullable: true
          readOnly: true
        notes:
          type: string
          nullable: true
          readOnly: true
        source:
          type: string
          enum: [application, workload, namespace, none]
          readOnly: true
...